move non fixtures out of fixtures section
