+++ tests/algorithms/test_factorize_algorithm.py
@@ -14,8 +14,8 @@ from composer.loggers import Logger
 from composer.models import HuggingFaceModel
 from composer.utils import module_surgery
 from tests.common import SimpleConvModel, SimpleTransformerClassifier
+from tests.common.datasets import dummy_tiny_bert_lm_batch, dummy_transformer_classifier_batch
 from tests.common.models import tiny_bert
-from tests.fixtures.inputs import dummy_tiny_bert_lm_batch, dummy_transformer_classifier_batch
 
 
 def create_state(minimal_state: State, model):
