+++ tests/algorithms/test_gradient_clipping.py
@@ -14,8 +14,8 @@ from composer.algorithms.gradient_clipping.gradient_clipping import _apply_agc, 
 from composer.core import Engine, State
 from composer.core.event import Event
 from tests.common import world_size
+from tests.common.datasets import dummy_tiny_bert_classification_batch, dummy_transformer_classifier_batch
 from tests.common.models import SimpleTransformerClassifier, configure_tiny_bert_config
-from tests.fixtures.inputs import dummy_tiny_bert_classification_batch, dummy_transformer_classifier_batch
 
 
 def simple_model_with_grads():
