+++ tests/common/datasets.py
@@ -4,10 +4,11 @@
 from typing import Sequence
 
 import torch
-import torch.utils.data
 from PIL import Image
-from torch.utils.data import Dataset
+from torch.utils.data import DataLoader, Dataset
 from torchvision.datasets import VisionDataset
+
+from composer.utils import dist
 
 
 class RandomClassificationDataset(Dataset):
@@ -122,7 +123,7 @@ class RandomSegmentationDataset(VisionDataset):
             return x, y
 
 
-class RandomTextClassificationDataset(torch.utils.data.Dataset):
+class RandomTextClassificationDataset(Dataset):
     """ Text classification dataset with values (just input token ids) drawn uniformly
     Args:
         vocab_size (int): vocab size to use (default: 10)
@@ -170,7 +171,7 @@ class RandomTextClassificationDataset(torch.utils.data.Dataset):
             return x, y
 
 
-class RandomTextLMDataset(torch.utils.data.Dataset):
+class RandomTextLMDataset(Dataset):
     """ Text LM dataset with values (just input token ids) drawn uniformly
     Args:
         vocab_size (int): vocab size to use (default: 10)
@@ -205,3 +206,49 @@ class RandomTextLMDataset(torch.utils.data.Dataset):
             return {'input_ids': x}
         else:
             return x
+
+
+def dummy_transformer_classifier_batch(vocab_size=100, num_classes=2):
+    sequence_length = 32
+    size = 100
+    batch_size = 8
+    train_dataset = RandomTextClassificationDataset(size=size,
+                                                    vocab_size=vocab_size,
+                                                    sequence_length=sequence_length,
+                                                    num_classes=num_classes)
+    train_dataloader = DataLoader(train_dataset, batch_size=batch_size, sampler=dist.get_sampler(train_dataset))
+    return next(iter(train_dataloader))
+
+
+def dummy_tiny_bert_classification_batch():
+    vocab_size = 30522  # Match bert vocab size
+    sequence_length = 4
+    num_classes = 2
+    size = 16
+    batch_size = 8
+
+    train_dataset = RandomTextClassificationDataset(size=size,
+                                                    vocab_size=vocab_size,
+                                                    sequence_length=sequence_length,
+                                                    num_classes=num_classes,
+                                                    use_keys=True)
+
+    train_dataloader = DataLoader(train_dataset, batch_size=batch_size, sampler=dist.get_sampler(train_dataset))
+    batch = next(iter(train_dataloader))
+    return batch
+
+
+def dummy_tiny_bert_lm_batch():
+    vocab_size = 30522  # Match bert vocab size
+    sequence_length = 4
+    size = 16
+    batch_size = 8
+
+    train_dataset = RandomTextLMDataset(size=size,
+                                        vocab_size=vocab_size,
+                                        sequence_length=sequence_length,
+                                        use_keys=True)
+
+    train_dataloader = DataLoader(train_dataset, batch_size=batch_size, sampler=dist.get_sampler(train_dataset))
+    batch = next(iter(train_dataloader))
+    return batch
